Revert merge PR #264
The changes in the above commit are not included in the DG submitted
to Canvas. It is past the tP final deliverables deadline. No other
files were changed in the above PR.

CS2103/T requires the repo to match the other final deliverables.

Let's revert the commit so that the repo matches our Canvas submission.

A revert is done instead of a reset to avoid re-writing history.

Prof Damith has given permission for this change to the repo, despite
the code-freeze period being active (see email sent on 4 Nov,
9:27 pm).